Corrected the image showing on bids
I missed showing the images of the freelancers and had some random text showing. Corrected that and now the images are getting shown. This will work on ly locally for me as image is stored locally on my folder. I did not found any other way to work this out as of now.